split ConnectionControler and ConnectionData
